Improved dataset handling.
Dataset handling is now more fault tolerant:
miniGAP won't terminate in error now when you request more structures
than there are in the import file.

Two new options for managing the dataset:
(1) You can choose the fraction of a dataset from a file to use in miniGAP.
Before you had to select the total number of structures.
(2) You can select the absolute number of structures in the training set.
Before you could only select a fraction of the imported dataset.